[DEV-284] refactor: 사용하지 않는 레포지토리 삭제 및 북마크 컨트롤러 추가
- BookmarkController 추가
- /v1/bookmark/** 유저만 요청 가능하도록 Security 추가
