+++ src/main/java/com/onedreamus/project/global/config/security/SecurityConfig.java
@@ -52,8 +52,8 @@ public class SecurityConfig {
         "/v1/user/logout", "/v1/note/**", "/v1/quiz/**", "/v1/users/quiz/first-attempt",
         "/v1/contents/news/{newsId}/users",
         "/v1/users/study-days/count",
-            "/v1/missions/**"
-
+        "/v1/missions/**",
+        "/v1/bookmark/**"
     };
 
     @Bean
