+++ src/main/java/com/onedreamus/project/thisismoney/repository/TagRepository.java
@@ -1,9 +1,10 @@
-package com.onedreamus.project.thisismoney.repository;
-
-import com.onedreamus.project.thisismoney.model.entity.Tag;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TagRepository extends JpaRepository<Tag, Integer> {
-}
+//package com.onedreamus.project.thisismoney.repository;
+//
+//import com.onedreamus.project.thisismoney.model.entity.Tag;
+//import org.springframework.data.jpa.repository.JpaRepository;
+//import org.springframework.stereotype.Repository;
+//
+//@Repository
+//public interface TagRepository extends JpaRepository<Tag, Integer> {
+//}
+//a
