+++ src/main/java/com/onedreamus/project/thisismoney/repository/TooltipRepository.java
@@ -1,9 +1,9 @@
-package com.onedreamus.project.thisismoney.repository;
-
-import com.onedreamus.project.thisismoney.model.entity.ScriptParagraphDictionary;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TooltipRepository extends JpaRepository<ScriptParagraphDictionary, Integer> {
-}
+//package com.onedreamus.project.thisismoney.repository;
+//
+//import com.onedreamus.project.thisismoney.model.entity.ScriptParagraphDictionary;
+//import org.springframework.data.jpa.repository.JpaRepository;
+//import org.springframework.stereotype.Repository;
+//
+//@Repository
+//public interface TooltipRepository extends JpaRepository<ScriptParagraphDictionary, Integer> {
+//}
